Create TextHighlighter component and remove react-highlight-words library
